search/gvg_browser/requirements.txt
# Minimal requirements to run GvG_Search_Function and its dependencies
openai>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
rich>=13.0.0
requests>=2.31.0
pgvector>=0.2.0
openpyxl>=3.1.0
reportlab>=4.0.0
dash>=2.17.1
dash-bootstrap-components>=1.6.0
supabase>=2.0.0
gunicorn>=21.2.0
rarfile>=4.1
stripe>=7.0.0
flask>=3.0.0

